refactor: Move test data creation in one class
On now already two points in the tests test data was created. For a
higher cohesion I created the class TestDataCreator, that has the single
responsibility to create test data. UserServiceTest and UserBuilderTest
are now calling this class to create test data for their tests instead
of creating them by themselves.
I didn't change the creation of data in UserRepository, because in a few
TDD cycles I will change this in reading from a json file either way
and I think a non test class should not use a class for test data
creation as a workaround.