refactor: Use assertEquals() on lists
I used a more complex structure of assertions (assert that list is not
null --> assert that size of list is correct --> for every item in list
assert that it is the expected item). But I realized that the assertion
if two lists are equal could be done much more efficient by using just
one assertEquals on the two lists. So I changed these parts of the code.