chore: Add Junit tags for integration tests
I needed an option to run only unit tests in Eclipse. One way to make
the test runner of Eclipse distinguish between unit and integration
tests is to give them different @Tag annotations.
So now all integration tests have the annotation @Tag("IntegrationTest")
and it works fine. Because of that I also adjusted failsafe- and
surefire-plugin in the Maven pom.