chore: Remove TODO for Mockito performance
There was a TODO comment because I experienced performance issues as I
implemented the first unit test und started to using Mockito in it.
The test took more than 400 ms to run. I thought and still think this is
not acceptable for a simple unit test and that's why I added the TODO to
check if I can improve the performance somehow.
Now that I have more than one test it seems that Mockito needs some time
to initialize. Because the first test that uses Mockito needs 400 to 900
ms to run and every test after the first one only takes 1 to 10 ms wich
is acceptable.