feat: Create first view in app component
I changed the app component, so that it calls userService to retrieve
all users from the backend and then display their names in the view.
The view is not finished but it is a simple prototype to see if
everthing works.

For the REST request I had to add an annotation to allow cross-origin
requests in the backend.