feat: Delete unit tests for UserRepository
I want to change the UserRepository to a Spring JPA Repository instead
of a JSON mapper and because the Spring JPA Repository has the crud
methods implemented out of the box, it makes no sense to unit test the
them. (see https://stackoverflow.com/a/23442457)
If I have some costum implementations I will have to create new unit
tests.